refactor: extract reef io crate
