+++ crates/reef-agent/Cargo.toml
@@ -12,8 +12,8 @@ name = "reef-agent"
 path = "src/main.rs"
 
 [dependencies]
-reef                = { path = "../.." }
 reef-core           = { path = "../reef-core" }
+reef-io             = { path = "../reef-io" }
 reef-proto          = { path = "../reef-proto" }
 reef-sqlite-preview = { path = "../reef-sqlite-preview" }
 serde               = { workspace = true }
