+++ crates/reef-io/src/agent_deploy/embedded.rs
