+++ src/lib.rs
@@ -2,16 +2,13 @@
 //! integration tests under `tests/` can exercise them. The `reef` binary
 //! (src/main.rs) is a thin wrapper that consumes this library.
 
-pub mod agent_deploy;
 pub mod app;
-pub mod backend;
 pub mod clipboard;
 pub mod editor;
 pub mod file_clipboard;
 pub mod file_selection;
 pub mod file_tree;
 pub mod find_widget;
-pub mod fs_watcher;
 pub mod global_search;
 pub mod graph_branch_picker;
 pub mod hosts_picker;
