+++ tests/backend_numstat_loopback.rs
@@ -8,7 +8,7 @@
 use std::path::Path;
 use std::sync::Mutex;
 
-use reef::backend::{Backend, LocalBackend, RemoteBackend};
+use reef_io::{Backend, LocalBackend, RemoteBackend};
 use test_support::{agent_bin, commit_file, tempdir_repo, write_file};
 
 static BACKEND_LOCK: Mutex<()> = Mutex::new(());
