+++ tests/backend_upload_loopback.rs
@@ -10,7 +10,7 @@
 use std::path::Path;
 use std::sync::Mutex;
 
-use reef::backend::{Backend, BackendError, LocalBackend, RemoteBackend};
+use reef_io::{Backend, BackendError, LocalBackend, RemoteBackend};
 use tempfile::TempDir;
 use test_support::agent_bin;
